+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This page summarizes historic changes in the library. Please also see the
 
 ## 0.3
 
+### 0.3.2
+
+- Fix for Python 3.10
+
 ### 0.3.1
 
 - `initialize` accepts nested dictionaries for `default`/`constant` arguments.